backend/pyproject.toml
[project]
name = "lead-sender-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "httpx>=0.27",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "google-genai>=0.3",
    "playwright>=1.44",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.1",
    "beautifulsoup4>=4.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
